make focusable-arrow-keys-behavior into a mixin
